requirements.txt
# local package
-e .

# external requirements
click~=7.0
Sphinx
coverage
awscli
flake8
python-dotenv~=0.14.0

odfpy~=1.4.1
python-docx~=0.8.10
pandas~=1.1.0
numpy~=1.19.0
google-api-python-client~=1.10.0
pdfminer3~=2018.12.3.0
gspread~=3.6.0
sklearn~=0.0
scikit-learn~=0.23.2
spacy~=2.3.2
Unidecode~=1.1.1
requests~=2.24.0
setuptools~=45.2.0